rm presets.py; use PresetRegistry in cli; mv presets.py functions to cli/registry
