+++ src/fetchez/api.py
@@ -27,7 +27,13 @@ from typing import List, Optional, Dict, Any
 from .utils import parse_hook_string
 from .core import run_fetchez
 from .spatial import parse_region
-from .registry import ModuleRegistry, HookRegistry, RecipeRegistry, SchemaRegistry, PresetRegistry
+from .registry import (
+    ModuleRegistry,
+    HookRegistry,
+    RecipeRegistry,
+    SchemaRegistry,
+    PresetRegistry,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -71,6 +77,7 @@ def list_recipes() -> Dict[str, Any]:
 
 def list_schemas() -> Dict[str, Any]:
     return _search_registry(SchemaRegistry)
+
 
 def list_presets() -> Dict[str, Any]:
     return _search_registry(PresetRegistry)
