+++ src/fetchez/registry.py
@@ -415,3 +415,23 @@ class PresetRegistry:
     @classmethod
     def get_preset(cls, name: str) -> Optional[Dict[str, Any]]:
         return cls.get_registry().get(name)
+
+    @classmethod
+    def hook_list_from_preset(cls, preset_def):
+        """Convert yaml definition to list of Hook Objects."""
+
+        hooks = []
+        for h_def in preset_def.get("hooks", []):
+            name = h_def.get("name")
+            kwargs = h_def.get("args", {})
+
+            hook_cls = HookRegistry.get_class(name)
+            if hook_cls:
+                try:
+                    hooks.append(hook_cls(**kwargs))
+                except Exception as exception:
+                    logger.error(f"Failed to init preset hook '{name}': {exception}")
+            else:
+                logger.warning(f"Preset hook '{name}' not found.")
+
+        return hooks
